Moving and renaming.
